requirements.txt
requests
beautifulsoup4
argparse
vercel-blob
python-dotenv